+++ TB2J/manager.py
@@ -32,6 +32,7 @@ def gen_exchange(path,
                  Rcut=None,
                  ne=None,
                  use_cache=False,
+                 output_path='TB2J_results',
                  description=''):
     atoms = read(os.path.join(path, posfile))
     basis_fname = os.path.join(path, 'basis.txt')
@@ -74,7 +75,7 @@ Warning: Please check if the noise level of Wannier function Hamiltonian to make
                                ne=ne,
                                use_cache=use_cache,
                                description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print(
             "All calculation finsihed. The results are in TB2J_results directory."
         )
@@ -119,7 +120,7 @@ Warning: Please check if the noise level of Wannier function Hamiltonian to make
                               ne=ne,
                               use_cache=use_cache,
                               description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print(
             "All calculation finsihed. The results are in TB2J_results directory."
         )
@@ -167,9 +168,9 @@ Warning: Please check if the noise level of Wannier function Hamiltonian to make
                                use_cache=use_cache,
                                description=description)
         print("\n")
-        exchange.run()
+        exchange.run(path=output_path)
         print(
-            "All calculation finsihed. The results are in TB2J_results directory."
+            f"All calculation finsihed. The results are in {output_path} directory."
         )
 
 
@@ -188,6 +189,7 @@ def gen_exchange_siesta(
     Rcut=None,
     ne=None,
     use_cache=False,
+    output_path='TB2J_results',
     description=''):
 
     try:
@@ -197,7 +199,7 @@ def gen_exchange_siesta(
             "sisl cannot be imported. Please install sisl first.")
     fdf = sisl.get_sile(fdf_fname)
     H = fdf.read_hamiltonian()
-    if H.spin.is_colinear and True:
+    if H.spin.is_colinear:
         print("Reading Siesta hamiltonian: colinear spin.")
         tbmodel_up = SislWrapper(H, spin=0)
         tbmodel_dn = SislWrapper(H, spin=1)
@@ -226,10 +228,10 @@ def gen_exchange_siesta(
             ne=ne,
             use_cache=use_cache,
             description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print("\n")
         print(
-            "All calculation finsihed. The results are in TB2J_results directory."
+            f"All calculation finsihed. The results are in {output_path} directory."
         )
 
     elif H.spin.is_colinear:
@@ -261,10 +263,10 @@ def gen_exchange_siesta(
             ne=ne,
             use_cache=use_cache,
             description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print("\n")
         print(
-            "All calculation finsihed. The results are in TB2J_results directory."
+            f"All calculation finsihed. The results are in {output_path} directory."
         )
 
     elif H.spin.is_spinorbit:
@@ -294,10 +296,10 @@ Warning: The DMI component parallel to the spin orientation, the Jani which has 
                                ne=ne,
                                use_cache=use_cache,
                                description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print("\n")
         print(
-            "All calculation finsihed. The results are in TB2J_results directory."
+            f"All calculation finsihed. The results are in {output_path} directory."
         )
 
 
@@ -310,6 +312,7 @@ def gen_exchange_gpaw(gpw_fname,
                       exclude_orbs=[],
                       Rcut=None,
                       use_cache=False,
+                      output_path='TB2J_results',
                       description=''):
     print("Reading from GPAW data and calculate electronic structure.")
     model = GPAWWrapper(gpw_fname=gpw_fname)
@@ -332,8 +335,8 @@ def gen_exchange_gpaw(gpw_fname,
                                Rcut=Rcut,
                                use_cache=use_cache,
                                description=description)
-        exchange.run()
+        exchange.run(path=output_path)
         print("\n")
         print(
-            "All calculation finsihed. The results are in TB2J_results directory."
+            f"All calculation finsihed. The results are in {output_path} directory."
         )